+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN6WVS8CJHZ5Z5REAVT0XVZ5"
+        "01KN91R2Q143306XDVXGF0RZSZ"
     )
     r"""The unique identifier of the data source"""