+++ src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN7132X1TS3YD3BZ0GECCBQK"
+        "01KN7YD6F3WY2YEBKKFQ03PFHM"
     )
     r"""The unique identifier of the data source"""