+++ src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN7132X4N1BEG8E52RJ916V2"
+        "01KN7YD6F6PSCXTVSPDKKAE7CE"
     )
     r"""The unique identifier of the data source"""