+++ packages/orq-rc/RELEASES.md
@@ -6829,3 +6829,13 @@ Based on:
 - [python v4.7.0-rc.43] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.43 - packages/orq-rc
+
+## 2026-04-03 06:53:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.44] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.44] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.44 - packages/orq-rc